refactor: move FixedPointTester
